dijkstra algorithm

+++ src/main/java/socs/network/message/LinkDescription.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class LinkDescription implements Serializable {
   public String linkID;
   public int portNum;
-  public int tosMetrics;
+  public short tosMetrics;
 
   public String toString() {
     return linkID + ","  + portNum + "," + tosMetrics;
